+++ tests/test_subtraction/test.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 
 set -e
+
+if [ -z "${SPECASM_TARGET_NEXT_OPCODES}" ]; then
+    exit 0
+fi
+
 rm subtraction 2>/dev/null 1>&2 || true
 
 ../../salink 2>/dev/null 1>&2
